Backend/model-service/scraper/requirements.txt
# Comment Scraper - YouTube & TikTok
# Tối ưu với async, retry, validation, NLP labeling (PhoBERT)

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# HTTP & ASYNC
# ============================================================================
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# ============================================================================
# YOUTUBE API
# ============================================================================
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# ============================================================================
# TIKTOK SCRAPING (Selenium-based)
# ============================================================================
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# ============================================================================
# NLP & ML - PhoBERT (Vietnamese NLP)
# ============================================================================
transformers>=4.30.0
torch>=2.0.0
# Vietnamese sentiment models (tự động download khi chạy):
# - vinai/phobert-base (~400MB)
# - wonrax/phobert-base-vietnamese-sentiment

# ============================================================================
# DATA PROCESSING
# ============================================================================
regex>=2023.0.0

# ============================================================================
# PROGRESS & LOGGING
# ============================================================================
tqdm>=4.66.0
colorlog>=6.7.0

# ============================================================================
# UTILITIES
# ============================================================================
tenacity>=8.2.0  # Advanced retry library
